+++ ui/app/src/components/ChangeContentTypeDialog/ChangeContentTypeDialogContainer.tsx
@@ -21,7 +21,6 @@ import React, { useEffect, useMemo, useState } from 'react';
 import { LegacyContentType } from '../../models/ContentType';
 import { fetchLegacyContentTypes } from '../../services/contentTypes';
 import { showErrorDialog } from '../../state/reducers/dialogs/error';
-import { useLogicResource } from '../../hooks/useLogicResource';
 import { useSubject } from '../../hooks/useSubject';
 import { debounceTime } from 'rxjs/operators';
 import DialogBody from '../DialogBody/DialogBody';
@@ -29,10 +28,10 @@ import { Box, Checkbox, FormControlLabel } from '@mui/material';
 import SingleItemSelector from '../SingleItemSelector';
 import { FormattedMessage } from 'react-intl';
 import SearchBar from '../SearchBar/SearchBar';
-import { SuspenseWithEmptyState } from '../Suspencified/Suspencified';
 import { ContentTypesGrid, ContentTypesLoader } from '../NewContentDialog';
 import DialogFooter from '../DialogFooter/DialogFooter';
 import { makeStyles } from 'tss-react/mui';
+import EmptyState from '../EmptyState';
 
 const useStyles = makeStyles()(() => ({
   compact: {
@@ -61,8 +60,13 @@ export function ChangeContentTypeDialogContainer(props: ChangeContentTypeDialogC
   const [openSelector, setOpenSelector] = useState(false);
   const [selectedItem, setSelectedItem] = useState(item);
   const [contentTypes, setContentTypes] = useState<LegacyContentType[]>();
+  const [isFetching, setIsFetching] = useState(false);
   const [keyword, setKeyword] = useState('');
   const [debounceKeyword, setDebounceKeyword] = useState('');
+  const filteredContentTypes = useMemo(() => {
+    const lowercaseKeyword = debounceKeyword.toLowerCase();
+    return contentTypes?.filter((contentType) => contentType.label.toLowerCase().includes(lowercaseKeyword));
+  }, [contentTypes, debounceKeyword]);
 
   const onSelectedContentType = (contentType: LegacyContentType) => {
     onContentTypeSelected?.({
@@ -72,8 +76,10 @@ export function ChangeContentTypeDialogContainer(props: ChangeContentTypeDialogC
 
   useEffect(() => {
     if (selectedItem.path) {
-      fetchLegacyContentTypes(site, selectedItem.path).subscribe(
-        (response) => {
+      setIsFetching(true);
+      const sub = fetchLegacyContentTypes(site, selectedItem.path).subscribe({
+        next: (response) => {
+          setIsFetching(false);
           setContentTypes(
             response.filter(
               (contentType) =>
@@ -81,27 +87,16 @@ export function ChangeContentTypeDialogContainer(props: ChangeContentTypeDialogC
             )
           );
         },
-        (response) => {
+        error: (response) => {
+          setIsFetching(false);
           dispatch(showErrorDialog({ error: response }));
         }
-      );
+      });
+      return () => {
+        sub.unsubscribe();
+      };
     }
   }, [dispatch, selectedItem, site]);
-
-  const resource = useLogicResource(
-    useMemo(() => ({ contentTypes, debounceKeyword }), [contentTypes, debounceKeyword]),
-    {
-      shouldResolve: ({ contentTypes }) => Boolean(contentTypes),
-      shouldReject: () => null,
-      shouldRenew: (source, resource) => resource.complete,
-      resultSelector: ({ contentTypes, debounceKeyword }) => {
-        return contentTypes.filter((contentType) =>
-          contentType.label.toLowerCase().includes(debounceKeyword.toLowerCase())
-        );
-      },
-      errorSelector: () => null
-    }
-  );
 
   const onSearch$ = useSubject<string>();
 
@@ -138,32 +133,32 @@ export function ChangeContentTypeDialogContainer(props: ChangeContentTypeDialogC
             <SearchBar onChange={onSearch} keyword={keyword} autoFocus showActionButton={Boolean(keyword)} />
           </Box>
         </Box>
-        <SuspenseWithEmptyState
-          resource={resource}
-          suspenseProps={{
-            fallback: <ContentTypesLoader numOfItems={6} isCompact={isCompact} />
-          }}
-          withEmptyStateProps={{
-            emptyStateProps: {
-              classes: {
-                image: classes.emptyStateImg
-              },
-              title: (
+        {isFetching ? (
+          <ContentTypesLoader numOfItems={6} isCompact={isCompact} />
+        ) : filteredContentTypes ? (
+          filteredContentTypes.length > 0 ? (
+            <ContentTypesGrid
+              contentTypes={filteredContentTypes}
+              isCompact={isCompact}
+              onTypeOpen={onSelectedContentType}
+              selectedContentType={selectedContentType}
+            />
+          ) : (
+            <EmptyState
+              title={
                 <FormattedMessage
                   id="changeContentTypeDialog.emptyStateMessage"
                   defaultMessage="No Content Types Found"
                 />
-              )
-            }
-          }}
-        >
-          <ContentTypesGrid
-            resource={resource}
-            isCompact={isCompact}
-            onTypeOpen={onSelectedContentType}
-            selectedContentType={selectedContentType}
-          />
-        </SuspenseWithEmptyState>
+              }
+              classes={{
+                image: classes.emptyStateImg
+              }}
+            />
+          )
+        ) : (
+          <></>
+        )}
       </DialogBody>
       <DialogFooter>
         <FormControlLabel