+++ ui/app/src/components/PreviewPageExplorerPanel/PreviewPageExplorerPanel.tsx
@@ -42,8 +42,6 @@ import {
   sortItemOperationComplete
 } from '../../state/actions/preview';
 import { getHostToGuestBus, getHostToHostBus } from '../../utils/subjects';
-import Suspencified from '../Suspencified/Suspencified';
-import { Resource } from '../../models/Resource';
 import palette from '../../styles/palette';
 import { useDispatch } from 'react-redux';
 import Typography from '@mui/material/Typography';
@@ -56,10 +54,10 @@ import { showItemMegaMenu } from '../../state/actions/dialogs';
 import { useSelection } from '../../hooks/useSelection';
 import { useActiveSiteId } from '../../hooks/useActiveSiteId';
 import { usePreviewGuest } from '../../hooks/usePreviewGuest';
-import { useLogicResource } from '../../hooks/useLogicResource';
 import { useUnmount } from '../../hooks/useUnmount';
 import { useSpreadState } from '../../hooks/useSpreadState';
 import { SimpleTreeView } from '@mui/x-tree-view';
+import { LoadingState } from '../LoadingState';
 
 const rootPrefix = '{root}_';
 
@@ -716,17 +714,6 @@ export function PreviewPageExplorerPanel() {
     setKeyword(keyword);
   };
 
-  const resource = useLogicResource<boolean, any>(
-    { models, byId: ContentTypesById },
-    {
-      shouldResolve: (source) => Boolean(source.models && source.byId),
-      shouldReject: () => false,
-      shouldRenew: () => !Object.keys(processedModels.current).length && resource.complete,
-      resultSelector: () => true,
-      errorSelector: null
-    }
-  );
-
   useUnmount(onBack);
 
   return (
@@ -753,7 +740,7 @@ export function PreviewPageExplorerPanel() {
           <SearchBar showActionButton={Boolean(keyword)} onChange={handleSearchKeyword} keyword={keyword} />
           <Divider className={classes.divider} />
         </div>
-        <Suspencified loadingStateProps={{ title: formatMessage(translations.loading) }}>
+        {models && ContentTypesById ? (
           <PageExplorerUI
             handleBreadCrumbClick={handleBreadCrumbClick}
             handleClick={handleClick}
@@ -762,21 +749,21 @@ export function PreviewPageExplorerPanel() {
             optionsMenu={optionsMenu}
             rootPrefix={rootPrefix}
             handleOptions={handleOptions}
-            resource={resource}
             keyword={keyword}
             nodeLookup={nodeLookup}
             selected={state.selected}
             breadcrumbs={state.breadcrumbs}
             rootChildren={Object.keys(processedModels.current)}
           />
-        </Suspencified>
+        ) : (
+          <LoadingState title={formatMessage(translations.loading)} />
+        )}
       </SimpleTreeView>
     </>
   );
 }
 
 interface PageExplorerUIProps {
-  resource: Resource<any>;
   optionsMenu: {
     modelId: string;
     anchorEl: Element;
@@ -797,7 +784,6 @@ interface PageExplorerUIProps {
 
 function PageExplorerUI(props: PageExplorerUIProps) {
   const {
-    resource,
     handleScroll,
     handleClick,
     handleOptions,
@@ -813,8 +799,6 @@ function PageExplorerUI(props: PageExplorerUIProps) {
   } = props;
   const { classes } = useStyles();
   const { formatMessage } = useIntl();
-
-  resource.read();
 
   let node: any = null;